refactor[be]: Refactoring service architecture
- 기존의 CRUD 디렉토리 구조에서 depencies를 활용한 서비스 클래스 구조로 변경
